Move ServicePort api serialization from `ipam` to `dcim`
