+++ netbox/ipam/api/views.py
@@ -1,6 +1,6 @@
 from rest_framework import generics
 
-from ipam.models import VRF, Role, RIR, Aggregate, Prefix, IPAddress, VLAN, VLANGroup, ServicePort
+from ipam.models import VRF, Role, RIR, Aggregate, Prefix, IPAddress, VLAN, VLANGroup
 from ipam import filters
 
 from extras.api.views import CustomFieldModelAPIView
@@ -133,26 +133,6 @@ class IPAddressDetailView(CustomFieldModelAPIView, generics.RetrieveAPIView):
     queryset = IPAddress.objects.select_related('vrf__tenant', 'tenant', 'interface__device', 'nat_inside')\
         .prefetch_related('nat_outside', 'custom_field_values__field')
     serializer_class = serializers.IPAddressSerializer
-
-
-#
-# Service Port
-#
-
-class ServicePortListView(generics.ListAPIView):
-    """
-    List IP addresses (filterable)
-    """
-    queryset = ServicePort.objects.select_related('device', 'ip_address', 'port', 'protocol', 'name', 'description')
-    serializer_class = serializers.ServicePortSerializer
-
-
-class ServicePortDetailView(generics.RetrieveAPIView):
-    """
-    Retrieve a single IP address
-    """
-    queryset = ServicePort.objects.select_related('device', 'ip_address', 'port', 'protocol', 'name', 'description')
-    serializer_class = serializers.ServicePortSerializer
 
 
 #
